Add lighting control
Why: After reviewing code for the Lamp object, seems that this is really
for heat control and not intended as a light source

* ...

Added defines for LIGHT_PIN intended to attach to a relay for control of
lighting circuits.

Added LIGHTa object for feedback.

Added lightStatus to global state to keep track of if the light is on or
off

* ...